fix: Client의 weight, height를 Double->Float로 변경

See merge request s12-final/S12P31S209!122

Former-commit-id: c203083e59c486ac22c59a0ee70912fed3ce0abe

+++ BE/src/main/java/com/S209/yobi/DTO/requestDTO/ClientRequestDTO.java
@@ -30,10 +30,10 @@ public class ClientRequestDTO {
     private Integer gender;
 
     @NotNull
-    private Double height;
+    private Float height;
 
     @NotNull
-    private Double weight;
+    private Float weight;
 
     @Size(max = 100)
     @NotNull
